llm/requirements.txt
langchain==0.1.0
fastapi==0.100.0
uvicorn==0.22.0
python-dotenv==1.0.0
pydantic==2.0.3
requests==2.31.0
langchain-community==0.0.11
gpt4all==0.1.7
chromadb==0.4.15
